+++ components/com_crowdfunding/views/project/tmpl/manager_basic.php
@@ -36,6 +36,18 @@ defined('_JEXEC') or die;
                 <td><?php echo JText::_('COM_CROWDFUNDING_RAISED'); ?></td>
                 <td><?php echo $this->raised;?></td>
             </tr>
+            <tr>
+                <td><?php echo JText::_('COM_CROWDFUNDING_STARTING_DATE'); ?></td>
+                <td><?php echo JHtml::_('date', $this->item->funding_start, JText::_('DATE_FORMAT_LC3'));?></td>
+            </tr>
+            <tr>
+                <td><?php echo JText::_('COM_CROWDFUNDING_ENDING_DATE'); ?></td>
+                <td><?php echo JHtml::_('date', $this->item->funding_end, JText::_('DATE_FORMAT_LC3'));?></td>
+            </tr>
+            <tr>
+                <td><?php echo JText::_('COM_CROWDFUNDING_APPROVED'); ?></td>
+                <td><?php echo JHtml::_('crowdfunding.approved', $this->item->approved); ?></td>
+            </tr>
         </tbody>
     </table>
 </div>