+++ administrator/components/com_crowdfunding/CHANGELOG.md
@@ -9,8 +9,8 @@ Crowdfunding Platform Release Notes
 * Added option to the module Crowdfunding Info. You can use them to change button label, hide funding type and information for fundraising process.
 * Added tips and tricks on dashboard.
 * Improved functionality to change the state of reward, selected during payment process. The reward is assigned to transaction.
+* The text in the payment process was altered to sound more universal.
 * Improved usability and user experience.
-   * Added colour to the rows of the transactions.
 
 ###v2.3.3
 * [HIGH] The system does not display rewards if there is no created types.